Revert "Added Maze Class.cs"
This reverts commit e69e75a3cecb610c17d2a48c943bc255a6cc8db7.
