--- VR Maze/Assets/Scripts/MazeClass.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 70e540a775b884a42ad2959868ff8482
-timeCreated: 1488134427
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
